Created a salary slip

+++ arithmetic.py
@@ -1,7 +1,7 @@
 #progrram to accept two numbers and perform arithmetic operations
 
-num1 = int(input("Enter first number: "))
-num2 = int(input("Enter second number: "))
+num1 = float(input("Enter first number: "))
+num2 = float(input("Enter second number: "))
 
 print("Addition: ", (num1 + num2))
 print("Subtraction: ", (num1 - num2))
